requirements.txt
# Updated requirements.txt - versions managed by pip
# Core web framework
fastapi
uvicorn
pydantic

# Database related
sqlalchemy
psycopg2-binary

# HTTP client and utilities
httpx
python-multipart
python-dotenv

# LangChain related packages
langchain
langchain-core
langchain-community
langchain-openai
langchain-experimental

# OpenAI and related
openai
tiktoken

# Additional recommended packages for production
pydantic-settings  # Better environment variable handling
structlog         # Better logging
prometheus-client # Metrics collection 